Fix MinGW build errors and warnings
- Fix multiple definition error for thread_local variables in hazard_pointer
  - Move thread_local definitions to .cpp file for MinGW
  - Keep inline definitions for other compilers
- Fix pragma region warnings in MinGW by wrapping with _MSC_VER checks
- Fix sign conversion warning in GetACP() call
- Fix shadowing warning in callback_job constructor

+++ sources/utilities/conversion/convert_string.cpp
@@ -272,7 +272,7 @@ namespace utility_module
 	auto convert_string::get_system_code_page() -> int
 	{
 #ifdef _WIN32
-		return GetACP();
+		return static_cast<int>(GetACP());
 #else
 		return 65001;
 #endif
